add pytestcov and renames import because change project directory
