+++ praktikum/burger.py
@@ -1,7 +1,7 @@
 from typing import List
 
-from praktikum.bun import Bun
-from praktikum.ingredient import Ingredient
+from prakticum.bun import Bun
+from prakticum.ingredient import Ingredient
 
 
 class Burger:
