+++ tests/test_burger.py
@@ -1,7 +1,8 @@
 
-from burger import Burger
 import pytest
 from unittest.mock import Mock
+
+from praktikum.burger import Burger
 from data import Data
