+++ tests/test_database.py
@@ -1,6 +1,5 @@
-
 from data import Data
-from database import Database
+from praktikum.database import Database
 
 
 class TestDatabase:
